requirements.txt
Flask
faiss-cpu
openai
sentence-transformers
numpy
python-dotenv
